refactor: extract csv generation services
