+++ src/modules/advertising/advertising.service.ts
@@ -26,59 +26,6 @@ export class AdvertisingService {
         return items.map((item) => new AdvertisingEntity(item));
     }
 
-    async findManyCsv(filters: FilterAdvertisingDto): Promise<string> {
-        const items = await this.findMany(filters);
-        const header = [
-            'id',
-            'campaignId',
-            'sku',
-            'date',
-            'type',
-            'clicks',
-            'toCart',
-            'avgBid',
-            'moneySpent',
-            'minBidCpo',
-            'minBidCpoTop',
-            'competitiveBid',
-            'weeklyBudget',
-            'createdAt',
-        ];
-
-        const rows = items.map((item) =>
-            [
-                item.id,
-                item.campaignId,
-                item.sku,
-                item.date,
-                item.type,
-                item.clicks,
-                item.toCart,
-                item.avgBid,
-                item.moneySpent,
-                item.minBidCpo,
-                item.minBidCpoTop,
-                item.competitiveBid,
-                item.weeklyBudget,
-                item.createdAt,
-            ]
-                .map((value) => {
-                    if (value instanceof Date) {
-                        return value.toISOString();
-                    }
-
-                    if (value === undefined || value === null) {
-                        return '';
-                    }
-
-                    return String(value);
-                })
-                .join(','),
-        );
-
-        return [header.join(','), ...rows].join('\n');
-    }
-
     async getStatisticsExpense(date) {
         const ads = await this.advertisingApiService.getStatisticsExpense({
             from: date + 'T21:00:00Z',
