+++ src/modules/unit/unit.controller.ts
@@ -1,10 +1,14 @@
 import { Controller, Get, Header, Query } from '@nestjs/common';
 import { UnitService } from './unit.service';
 import { AggregateUnitDto } from './dto/aggregate-unit.dto';
+import { UnitCsvService } from '@/modules/unit/services/unit-csv.service';
 
 @Controller('unit')
 export class UnitController {
-  constructor(private readonly unitService: UnitService) {}
+  constructor(
+    private readonly unitService: UnitService,
+    private readonly unitCsvService: UnitCsvService,
+  ) {}
 
   @Get()
   aggregate(@Query() dto: AggregateUnitDto) {
@@ -15,6 +19,6 @@ export class UnitController {
   @Header('Content-Type', 'text/csv')
   @Header('Content-Disposition', 'attachment; filename="unit.csv"')
   aggregateCsv(@Query() dto: AggregateUnitDto) {
-    return this.unitService.aggregateCsv(dto);
+    return this.unitCsvService.aggregateCsv(dto);
   }
 }
